+++ app/controllers/blogpost_controller.js
@@ -1,3 +1,15 @@
+var controller = {}
+  , app
+  , db
+
+// Constructor
+
+module.exports = function (_app) {
+  app = _app
+  db  = app.set('db')
+  return controller
+}
+
 /**
  * Index BlogPost
  *
@@ -11,7 +23,7 @@
  * @url /posts
  */
 
-exports.index = function(req, res, db, next){
+controller.index = function(req, res, next){
   res.render('home', {
     posts: db.posts.getLatestPosts()
   });
@@ -29,7 +41,7 @@ exports.index = function(req, res, db, next){
  * @url /create
  */
 
-exports.create = function(req, res, db, next){
+controller.create = function(req, res, next){
   
   var BlogPost = db.main.model('BlogPost');
   var _post = new BlogPost(req.param('post'));